feat(main): modified to work with `config.Title` as string
If `config.Window = false` reports an error for invalid usage.

If `config.Window = true` if `config.Title` provided and diferent from
`"auto"` use it as title.

If `config.Window = true` and `config.Title` not provided or is
`"auto"`, use filename as title.

If `config.Window = true` and `config.Title` is empty string, don't set
title.